Move nest to util
