--- src/app/councillors/[contactSlug]/types/index.ts
@@ -1,1 +1,0 @@
-export * from '@/app/councillors/[contactSlug]/types';
