+++ src/database/queries/councillorVotes.ts
@@ -3,46 +3,21 @@ import { queryAndTagsToPostgresTextSearchQuery } from '@/logic/parseQuery';
 import { SearchPagination } from '@/logic/search';
 import { Kysely, sql } from 'kysely';
 
-type SearchAgendaItemArgs = {
-  options: { textQuery: string; contactSlug: string };
-  pagination: SearchPagination;
-};
+import { AgendaItem, Motion } from '@/app/councillors/[contactSlug]/types';
+
+type SearchOptions = { textQuery: string; contactSlug: string };
 
 export const searchCouncillorVotes = async (
   db: Kysely<DB>,
   {
-    options: { textQuery, contactSlug },
+    options,
     pagination: { page, pageSize },
-  }: SearchAgendaItemArgs,
+  }: {
+    options: SearchOptions;
+    pagination: SearchPagination;
+  },
 ) => {
-  const fuzzyQuery = queryAndTagsToPostgresTextSearchQuery({
-    textQuery,
-    tags: [],
-  });
-
-  if (!fuzzyQuery) throw new Error(`No query to run`);
-
-  const baseQuery = db
-    .with('AvailableConsiderations', (db) =>
-      db
-        .selectFrom('RawAgendaItemConsiderations')
-        .innerJoin('Votes', (eb) =>
-          eb.onRef('reference', '=', 'agendaItemNumber'),
-        )
-        .select(['textSearchVector', 'agendaItemNumber'])
-        .where('contactSlug', '=', contactSlug)
-        .distinct(),
-    )
-    .with('FuzzyQuery', (db) =>
-      db.selectNoFrom(sql`to_tsquery('english', ${fuzzyQuery})`.as('query')),
-    )
-    .with('SearchResults', (db) =>
-      db
-        .selectFrom(['AvailableConsiderations', 'FuzzyQuery'])
-        .select('agendaItemNumber')
-        .select(sql<number>`ts_rank("textSearchVector", query)`.as('rank'))
-        .whereRef('textSearchVector', '@@', 'query'),
-    );
+  const baseQuery = buildCouncillorVotesSearchResultsQuery(db, options);
 
   const totalCount = (
     await baseQuery
@@ -67,3 +42,60 @@ export const searchCouncillorVotes = async (
     results: rawResults,
   };
 };
+
+export function buildCouncillorVotesSearchResultsQuery(
+  db: Kysely<DB>,
+  options: SearchOptions,
+) {
+  const fuzzyQuery = queryAndTagsToPostgresTextSearchQuery({
+    textQuery: options.textQuery,
+    tags: [],
+  });
+  if (!fuzzyQuery) throw new Error(`No query to run`);
+  return db
+    .with('AvailableConsiderations', (db) =>
+      db
+        .selectFrom('RawAgendaItemConsiderations')
+        .innerJoin('Votes', (eb) =>
+          eb.onRef('reference', '=', 'agendaItemNumber'),
+        )
+        .select(['textSearchVector', 'agendaItemNumber'])
+        .where('contactSlug', '=', options.contactSlug)
+        .distinct(),
+    )
+    .with('FuzzyQuery', (db) =>
+      db.selectNoFrom(sql`to_tsquery('english', ${fuzzyQuery})`.as('query')),
+    )
+    .with('SearchResults', (db) =>
+      db
+        .selectFrom(['AvailableConsiderations', 'FuzzyQuery'])
+        .select('agendaItemNumber')
+        .select(sql<number>`ts_rank("textSearchVector", query)`.as('rank'))
+        .whereRef('textSearchVector', '@@', 'query'),
+    );
+}
+
+type FlatRow = Omit<AgendaItem & Motion, 'motions'>;
+export function nestMotionsUnderAgendaItems(
+  flatRows: ReadonlyArray<FlatRow>,
+): AgendaItem[] {
+  const agendaItemByNumber = new Map<string, AgendaItem>();
+  for (const {
+    agendaItemNumber,
+    agendaItemTitle,
+    agendaItemSummary,
+    aiSummary,
+    ...motion
+  } of flatRows) {
+    const agendaItem: AgendaItem = agendaItemByNumber.get(agendaItemNumber) ?? {
+      agendaItemNumber,
+      agendaItemTitle,
+      agendaItemSummary,
+      aiSummary,
+      motions: [],
+    };
+    agendaItem.motions.push(motion);
+    agendaItemByNumber.set(agendaItemNumber, agendaItem);
+  }
+  return [...agendaItemByNumber.values()];
+}
